bigquery-service/requirements.txt
flask>=2.0.2,<3.0
google-cloud-bigquery>=2.32.0
gunicorn>=20.1.0
